# transacoes/src/swagger/transacoes.yaml
openapi: 3.0.3
info:
  title: API Transações
  description: |-
    API do microsserviço de Transações
  version: 0.0.1
...
servers:
  - url: http://localhost:3002
paths:
...
  /api/admin/transactions/{id}:
    get:
      tags:
        - transactions
      description: get a transaction by id
      parameters:
        - name: id
          in: path
          description: ID of transaction that needs to be fetched
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfull operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
    patch:
      tags:
        - transactions
      description: change status of a transaction
      parameters:
        - name: id
          in: path
          description: ID of transaction that needs to be fetched
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum:
                    - 'Aprovada'
                    - 'Rejeitada'
      responses:
        '204':
          description: Success.
security:
  - auth: []
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        valor:
          type: number
        idUser:
          type: string
        status:
          type: string
          example: 'Em análise'
          enum:
            - 'Em análise'
            - 'Aprovada'
            - 'Reprovada'
  securitySchemes:
    auth:
      type: http
      scheme: bearer
      bearerFormat: JWT